refactor(ui): extract crossterm input-reader thread to ui/input_reader.rs (dirge-4y4l stage 12a)
First safe-stage extraction from the run_interactive god-function
(~3550-line async fn). Moves the blocking OS thread that polls crossterm
and forwards key/mouse/paste/resize events to the UI loop as UserEvents.

- ui/input_reader.rs (116 LOC) — spawn_input_reader(user_tx): the
  poll/read loop with cooperative shutdown via EVENT_READER_SHUTDOWN /
  EVENT_READER_EXITED. Self-contained: captures only a Sender<UserEvent>
  clone (caller keeps its own sender for other event sources).
- ui/mod.rs (3712 -> 3614): the ~97-line inline std::thread::spawn block
  becomes one call; drops the now-unused `crossterm::event` +
  MouseButton/MouseEventKind imports (only the reader used them).

Behavior-preserving. -D warnings clean default + windows-default;
2387 tests green.

Refs dirge-0b0s
