+++ .beads/issues.jsonl
@@ -66,6 +66,7 @@
 {"_type":"issue","id":"dirge-86e","title":"ANSI injection in permission ALERT prompt","description":"ask_req.tool / ask_req.input rendered un-sanitized at mod.rs:2584-2585. Reopen path already sanitizes — asymmetric. Sec impl: ANSI at the permission-decision moment.","status":"closed","priority":1,"issue_type":"bug","assignee":"Yogthos","owner":"yogthos@gmail.com","created_at":"2026-05-21T22:17:34Z","created_by":"Yogthos","updated_at":"2026-05-21T22:26:37Z","started_at":"2026-05-21T22:17:42Z","closed_at":"2026-05-21T22:26:37Z","close_reason":"Closed","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"dirge-9f1","title":"Chat history ignores 120-col content_width cap","description":"max_line_width and wrap_line use raw content_cols, so on wide terminals scrollback overflows the centered band into divider/panel margin.","status":"closed","priority":1,"issue_type":"bug","assignee":"Yogthos","owner":"yogthos@gmail.com","created_at":"2026-05-21T22:17:33Z","created_by":"Yogthos","updated_at":"2026-05-21T22:26:36Z","started_at":"2026-05-21T22:17:42Z","closed_at":"2026-05-21T22:26:36Z","close_reason":"Closed","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"dirge-woq","title":"R1: fix 3 critical plugin bugs (FFI panic, dialog deadlock, init hang)","description":"From the plugin subsystem audit: (1) wrap JanetCFunctions in catch_unwind so Rust panics don't unwind across the C-FFI boundary into Janet; (2) cancel send_dialog's reply_rx.recv() on worker shutdown so the worker thread doesn't block forever when the UI exits mid-dialog; (3) add timeout to the init handshake so a worker panic before init_tx.send() doesn't hang the main thread. Also: (4) bounds-assert wrap_string's i32 cast for the unlikely \u003e2GB case, (5) make take_string_slot atomic to close the race window, (6) don't eat unrelated user events in the dialog arm.","status":"closed","priority":1,"issue_type":"bug","owner":"yogthos@gmail.com","created_at":"2026-05-20T14:59:57Z","created_by":"Yogthos","updated_at":"2026-05-20T15:30:28Z","started_at":"2026-05-20T15:00:10Z","closed_at":"2026-05-20T15:30:28Z","dependency_count":0,"dependent_count":1,"comment_count":0}
+{"_type":"issue","id":"dirge-0b0s","title":"run_interactive stage 12a: extract crossterm input-reader thread to ui/input_reader.rs","description":"run_interactive in ui/mod.rs is a ~3550-line async fn (god-function; only its AgentEvent arms were extracted in M1). Safe-stage decomposition (user-approved scope). 12a: extract the crossterm input-reader thread (the std::thread::spawn loop that polls event::read and maps Key/Mouse/Paste/Resize -\u003e UserEvent, honoring EVENT_READER_SHUTDOWN/EXITED) into ui/input_reader.rs as spawn_input_reader(user_tx). Self-contained: captures only a Sender\u003cUserEvent\u003e clone. Behavior-preserving; -D warnings clean default+windows-default; full suite green. Follow-ups: 12b permission-decision sub-loop, 12c modified-files picker sub-loop. High-risk 950-line key-handling left inline by design.","status":"open","priority":2,"issue_type":"task","owner":"yogthos@gmail.com","created_at":"2026-06-01T22:09:03Z","created_by":"Yogthos","updated_at":"2026-06-01T22:09:03Z","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"dirge-gwpi","title":"syntax_validator: paren feedback missing for grammarless lisps (Janet etc.)","description":"check_syntax returns Ok(()) when no tree-sitter grammar exists for the extension, so the delimiter-balance scanner (which emits the actionable 'add N matching )' / 'do not count by hand' hint) never runs for .janet/.fnl/.lisp/.scm/.rkt/.el/.cljd/.jdn. The model writing Janet gets zero paren feedback and falls back to manual counting. Fix: (1) check_syntax falls back to the delimiter scanner when no grammar but lex rules exist; (2) add RULES_JANET (# line comments, backtick long-strings) since RULES_LISP's ';' comment char is wrong for Janet and would cause false positives; (3) format_errors uses accurate wording for the no-grammar path. TDD.","status":"closed","priority":2,"issue_type":"bug","owner":"yogthos@gmail.com","created_at":"2026-06-01T21:52:53Z","created_by":"Yogthos","updated_at":"2026-06-01T22:03:33Z","closed_at":"2026-06-01T22:03:33Z","close_reason":"Merged: PR #308","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"dirge-1g9q","title":"builder.rs stage 11c: extract build_agent_inner to agent_inner.rs","description":"Follow-up to dirge-553f/11b. Extract build_agent_inner (~516 LOC god-function) + hookify into agent_inner.rs, leaving builder/mod.rs as imports + module wiring + build_session_search_tool. Re-export via 'pub use'. build_agent_inner is generic \u003cM: CompletionModel\u003e and pub; keep the path crate::agent::builder::build_agent_inner valid. Heavy DI signature + many tool constructions; behavior-preserving move only (do NOT refactor the function body). -D warnings clean default+windows-default; full suite green.","notes":"Code complete in PR #307 (off main). agent_inner.rs (570) = build_agent_inner + hookify. mod.rs 586-\u003e33 (wiring + build_session_search_tool). reminder_tests.rs gains explicit Cli/Config/ToolCache/Sandbox/resolve_family imports. ALSO fixed 11b stray: restored wrap_mcp_tools #[cfg(feature=mcp)] + doc in loop_tools.rs, removed orphan that wrongly gated reminder_tests on mcp. End: builder.rs 1835 -\u003e builder/{mod 33, agent_inner 570, loop_tools 519, preamble 89, reminder_tests 697}. -D warnings clean default+windows-default; 2374 tests green. Leave in_progress until merge.","status":"closed","priority":2,"issue_type":"task","assignee":"Yogthos","owner":"yogthos@gmail.com","created_at":"2026-06-01T21:07:43Z","created_by":"Yogthos","updated_at":"2026-06-01T21:36:28Z","started_at":"2026-06-01T21:21:22Z","closed_at":"2026-06-01T21:36:28Z","close_reason":"Merged: #305 (11a), #306 (11b), #307 (11c)","dependencies":[{"issue_id":"dirge-1g9q","depends_on_id":"dirge-w11r","type":"blocks","created_at":"2026-06-01T17:07:50Z","created_by":"Yogthos","metadata":"{}"}],"dependency_count":1,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"dirge-w11r","title":"builder.rs stage 11b: extract build_loop_tools cluster to loop_tools.rs","description":"Follow-up to dirge-553f (11a, PR #305). Extract the LoopTool-registry builder cluster from builder/mod.rs into loop_tools.rs: build_loop_tools (~425 LOC god-function), wrap_mcp_tools, DynamicToolSearch struct, and build_session_search_tool + hookify if they belong with it. Re-export via 'pub use loop_tools::*' (build_loop_tools/wrap_mcp_tools/DynamicToolSearch are pub) so external paths unchanged. Mind the heavy import surface (all the tool types, mcp/semantic cfg-gates). reminder_tests 'use super::*' must keep resolving build_loop_tools/wrap_mcp_tools. -D warnings clean default+windows-default; full suite green.","notes":"Code complete in PR #306 (off main). loop_tools.rs (512) = build_loop_tools + wrap_mcp_tools + DynamicToolSearch, re-exported via 'pub use loop_tools::*'. Calls super::build_session_search_tool. mod.rs 1068-\u003e586 (build_agent_inner + hookify + wiring). -D warnings clean default+windows-default; 2374 tests green. Next: 11c (dirge-1g9q) build_agent_inner. Leave in_progress until merge.","status":"closed","priority":2,"issue_type":"task","assignee":"Yogthos","owner":"yogthos@gmail.com","created_at":"2026-06-01T21:07:42Z","created_by":"Yogthos","updated_at":"2026-06-01T21:36:28Z","started_at":"2026-06-01T21:10:58Z","closed_at":"2026-06-01T21:36:28Z","close_reason":"Merged: #305 (11a), #306 (11b), #307 (11c)","dependencies":[{"issue_id":"dirge-w11r","depends_on_id":"dirge-553f","type":"blocks","created_at":"2026-06-01T17:07:49Z","created_by":"Yogthos","metadata":"{}"}],"dependency_count":1,"dependent_count":1,"comment_count":0}
